vscbackend/src/requirements.txt
flask==2.0.3
bcrypt==4.0.1
cassandra-driver==3.29.0
streamlit==1.12.2
st-cookies-manager==0.2.2
pyjwt==2.4.0
requests==2.28.1
werkzeug>=2.0.0,<2.3.0
python-dotenv==0.19.2
groq==0.25.0
